Move adapters back
