+++ pkg/registry/core/adapters/nsm_client_to_server.go
@@ -17,13 +17,48 @@
 package adapters
 
 import (
-	"github.com/networkservicemesh/api/pkg/api/registry"
+	"context"
 
-	"github.com/networkservicemesh/sdk/pkg/registry/core/next"
-	"github.com/networkservicemesh/sdk/pkg/registry/core/next/adapters"
+	"github.com/golang/protobuf/ptypes/empty"
+	"github.com/networkservicemesh/api/pkg/api/registry"
 )
 
 // NewNSMClientToServer - returns a registry.NsmRegistryServer wrapped around the supplied client
 func NewNSMClientToServer(client registry.NsmRegistryClient) registry.NsmRegistryServer {
-	return adapters.NewNSMClientToServer(client, next.NSMRegistryServer)
+	return NewNSMClientToServerNext(client, nil)
 }
+
+type nsmClientToServer struct {
+	client registry.NsmRegistryClient
+	next   func(ctx context.Context) registry.NsmRegistryServer
+}
+
+// NewNSMClientToServerNext - returns a registry.NsmRegistryServer wrapped around the supplied client
+func NewNSMClientToServerNext(client registry.NsmRegistryClient, next func(ctx context.Context) registry.NsmRegistryServer) registry.NsmRegistryServer {
+	return &nsmClientToServer{client: client, next: next}
+}
+
+func (n *nsmClientToServer) RegisterNSM(ctx context.Context, r *registry.NetworkServiceManager) (*registry.NetworkServiceManager, error) {
+	result, err := n.client.RegisterNSM(ctx, r)
+	if err != nil || n.next == nil {
+		return result, err
+	}
+	return n.next(ctx).RegisterNSM(ctx, result)
+}
+
+func (n *nsmClientToServer) GetEndpoints(ctx context.Context, e *empty.Empty) (*registry.NetworkServiceEndpointList, error) {
+	result, err := n.client.GetEndpoints(ctx, e)
+	if err != nil || n.next == nil {
+		return nil, err
+	}
+	var nextResult *registry.NetworkServiceEndpointList
+	nextResult, err = n.next(ctx).GetEndpoints(ctx, e)
+	if err != nil {
+		return nil, err
+	}
+	result.NetworkServiceEndpoints = append(result.NetworkServiceEndpoints, nextResult.NetworkServiceEndpoints...)
+	return result, nil
+}
+
+// Implementation check
+var _ registry.NsmRegistryServer = &nsmClientToServer{}
