+++ pkg/registry/core/adapters/nsm_server_to_client.go
@@ -17,13 +17,49 @@
 package adapters
 
 import (
-	"github.com/networkservicemesh/api/pkg/api/registry"
+	"context"
 
-	"github.com/networkservicemesh/sdk/pkg/registry/core/next"
-	"github.com/networkservicemesh/sdk/pkg/registry/core/next/adapters"
+	"github.com/golang/protobuf/ptypes/empty"
+	"github.com/networkservicemesh/api/pkg/api/registry"
+	"google.golang.org/grpc"
 )
 
 // NewNSMServerToClient - returns a registry.NsmRegistryClient wrapped around the supplied server
 func NewNSMServerToClient(server registry.NsmRegistryServer) registry.NsmRegistryClient {
-	return adapters.NewNSMServerToClient(server, next.NSMRegistryClient)
+	return NewNSMServerToClientNext(server, nil)
 }
+
+type nsmServerToClient struct {
+	server registry.NsmRegistryServer
+	next   func(ctx context.Context) registry.NsmRegistryClient
+}
+
+// NewNSMServerToClientNext - returns a registry.NsmRegistryClient wrapped around the supplied server
+func NewNSMServerToClientNext(server registry.NsmRegistryServer, next func(ctx context.Context) registry.NsmRegistryClient) registry.NsmRegistryClient {
+	return &nsmServerToClient{server: server, next: next}
+}
+
+func (n *nsmServerToClient) RegisterNSM(ctx context.Context, in *registry.NetworkServiceManager, opts ...grpc.CallOption) (*registry.NetworkServiceManager, error) {
+	result, err := n.server.RegisterNSM(ctx, in)
+	if err != nil || n.next == nil {
+		return result, err
+	}
+	return n.next(ctx).RegisterNSM(ctx, in, opts...)
+}
+
+func (n *nsmServerToClient) GetEndpoints(ctx context.Context, in *empty.Empty, opts ...grpc.CallOption) (*registry.NetworkServiceEndpointList, error) {
+	result, err := n.server.GetEndpoints(ctx, in)
+	if err != nil || n.next == nil {
+		return nil, err
+	}
+	var nextResult *registry.NetworkServiceEndpointList
+	nextResult, err = n.next(ctx).GetEndpoints(ctx, in, opts...)
+	if err != nil {
+		return nil, err
+	}
+	result.NetworkServiceEndpoints = append(result.NetworkServiceEndpoints, nextResult.NetworkServiceEndpoints...)
+	return result, nil
+}
+
+// Implementation check
+var _ registry.NsmRegistryClient = &nsmServerToClient{}
